src/mlops_m6_project/monitoring_exercise_files/pyproject.toml
[project]
name = "monitoring-exercise-files"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "evidently>=0.7.20",
    "fastapi>=0.115.6",
    "google-cloud-storage>=3.7.0",
    "nltk>=3.9.2",
    "pydantic>=2.12.5",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
